+++ cypress/integration/main/LocalStorage.spec.ts
@@ -12,8 +12,14 @@ describe('Local Storage', () => {
         '"typeModal":null,"rulesets":[{"text": "Nuzlocke", "value":"Nuzlocke"}, {"text":"Test","value":"Test"}]'
       );
       const addFilter = addRuleset.replace('"pokemon":null', '"pokemon":null,"filter":["Test"]');
-      const replaceSoulocke = addFilter.replace('Soulocke', 'SomethingElse');
-      const changeVersion = replaceSoulocke.replace('"version":6', '"version":0');
+      const replaceShiny = addFilter.replace(
+        '"status":null',
+        '"status":{"value":6}, "details":{ "id":1,"level":1,"moves":[]}'
+      );
+      const replaceSoulocke = replaceShiny.replace('Soulocke', 'SomethingElse');
+      const replaceWedlocke = replaceSoulocke.replace('Wedlocke', 'SomethingElse');
+      const replaceBadge = replaceWedlocke.replace('"badge":[]', '"badge":null');
+      const changeVersion = replaceBadge.replace('"version":7', '"version":0');
       cy.setLocalStorage('pokemon-tracker', changeVersion);
     });
     cy.visit('/');
@@ -23,8 +29,9 @@ describe('Local Storage', () => {
   it('Remake Brilliant Diamond and Shining Pearl', () => {
     cy.visit('/');
     cy.getLocalStorage('pokemon-tracker').then((data) => {
-      const removeBDSP = data.replace('13.1', '15');
-      cy.setLocalStorage('pokemon-tracker', removeBDSP);
+      const removeBDSP = data.replace(/13.1/gi, '15');
+      const changeVersion = removeBDSP.replace('"version":7', '"version":4');
+      cy.setLocalStorage('pokemon-tracker', changeVersion);
     });
     cy.visit('/');
     cy.contains('Nuzlocke Tracker').should('exist');