+++ src/components/Tracker/elements/Encounters/Encounters.tsx
@@ -5,15 +5,28 @@ import { FixedSizeList, ListChildComponentProps as RowProps } from 'react-window
 import Confirm from 'semantic-ui-react/dist/commonjs/addons/Confirm';
 import Button from 'semantic-ui-react/dist/commonjs/elements/Button';
 import Icon from 'semantic-ui-react/dist/commonjs/elements/Icon';
+
 import Popup from 'semantic-ui-react/dist/commonjs/modules/Popup';
 import shallow from 'zustand/shallow';
 import { Status } from 'components';
-import { Detail, Evolve, Nickname, Pokemon, ScrollList, Swap } from 'components/Tracker/elements';
+import {
+  Detail,
+  Evolve,
+  Nature,
+  Nickname,
+  Pokemon,
+  ScrollList,
+  Swap,
+} from 'components/Tracker/elements';
 import { TYPE_COLOR } from 'constants/colors';
 import { POKEMAP } from 'constants/pokemon';
+import useRemtoPx from 'hooks/useRemToPx';
 import useStore from 'store';
 import { ReactComponent as PokeballSVG } from 'assets/svg/pokeball.svg';
 import styles from './Encounters.module.scss';
+
+const NICKNAME_HEIGHT = 15.714; // 220px
+const NORMAL_HEIGHT = 12.857; // 180px
 
 const Encounters = React.memo(function Encounters() {
   const { t } = useTranslation('tracker');
@@ -43,6 +56,7 @@ const Encounters = React.memo(function Encounters() {
   );
   const [encounterToDelete, setEncounterToDelete] = useState<number>(null);
   const [confirm, setConfirm] = useState(false);
+  const itemSize = useRemtoPx(nicknames ? NICKNAME_HEIGHT : NORMAL_HEIGHT);
 
   const filteredEncounters = useMemo(() => {
     return games[selectedGame?.value]?.encounters?.filter((enc) => {
@@ -172,6 +186,7 @@ const Encounters = React.memo(function Encounters() {
             {!!foundPokemon && [1, 3, 4, 7].includes(encounter.status?.value) && (
               <Swap encounter={encounter} />
             )}
+            {!!foundPokemon && <Nature encounter={encounter} />}
           </div>
         </div>
       </div>
@@ -186,7 +201,7 @@ const Encounters = React.memo(function Encounters() {
             <FixedSizeList
               height={690}
               itemCount={filteredEncounters?.length}
-              itemSize={nicknames ? 220 : 180}
+              itemSize={itemSize}
               ref={listRef}
               width="100%"
             >