+++ src/components/Tracker/elements/index.ts
@@ -3,6 +3,7 @@ export { default as Encounters } from './Encounters/Encounters';
 export { default as Detail } from './Detail/Detail';
 export { default as Evolve } from './Evolve/Evolve';
 export { default as FAB } from './FAB/FAB';
+export { default as Nature } from './Nature/Nature';
 export { default as Nickname } from './Nickname/Nickname';
 export { default as Options } from './Options/Options';
 export { default as Pokemon } from './Pokemon/Pokemon';